requirements.txt
requests
markdown
pillow
django
python-dotenv
gunicorn
psycopg2-binary
raven
django
djangorestframework
